Add fast clustering to existing file
