--- mteb/tasks/Clustering/deu/TenKGnadClusteringS2SFast.py
@@ -1,40 +1,0 @@
-from __future__ import annotations
-
-from mteb.abstasks.AbsTaskClusteringFast import AbsTaskClusteringFast, convert_to_fast
-from mteb.abstasks.TaskMetadata import TaskMetadata
-
-
-class TenKGnadClusteringS2S(AbsTaskClusteringFast):
-    metadata = TaskMetadata(
-        name="TenKGnadClusteringS2S.v2",
-        description="Clustering of news article titles. Clustering of 10 splits on the news article category.",
-        reference="https://tblock.github.io/10kGNAD/",
-        dataset={
-            "path": "slvnwhrl/tenkgnad-clustering-s2s",
-            "revision": "6cddbe003f12b9b140aec477b583ac4191f01786",
-        },
-        type="Clustering",
-        category="s2s",
-        eval_splits=["test"],
-        eval_langs=["deu-Latn"],
-        main_score="v_measure",
-        date=(
-            "2000-01-01",
-            "2020-12-31",
-        ),  # since it is news it is guessed that it is from 2000 to 2020
-        form=["written"],
-        domains=["News", "Non-fiction"],
-        task_subtypes=None,
-        license="cc-by-sa-4.0",
-        socioeconomic_status="medium",
-        annotations_creators="derived",
-        dialect=[],
-        text_creation="found",
-        bibtex_citation=None,  # none found
-        n_samples={"test": 10275},  # due to duplicates
-        avg_character_length={"test": 50.96},
-    )
-
-    def dataset_transform(self) -> None:
-        ds = convert_to_fast(self.dataset, self.seed)  # type: ignore
-        self.dataset = ds
