+++ results/sentence-transformers__paraphrase-multilingual-MiniLM-L12-v2/ArxivClusteringP2P.v2.json
@@ -1,21 +1,53 @@
 {
   "dataset_revision": "a122ad7f3f0291bf49cc6f4d32aa80929df69d5d",
-  "mteb_dataset_name": "ArxivClusteringP2P.v2",
-  "mteb_version": "1.6.37",
-  "test": {
-    "evaluation_time": 186.9,
-    "v_measure": 0.3866478037243582,
-    "v_measures": [
-      0.37289092097683685,
-      0.3914427448155115,
-      0.3713691481207848,
-      0.3793698549707203,
-      0.39326714936605683,
-      0.39451528831035443,
-      0.40487469284188676,
-      0.39423765628184715,
-      0.36103812817103303,
-      0.40347245338855003
+  "evaluation_time": 399.76936388015747,
+  "kg_co2_emissions": null,
+  "mteb_version": "1.11.19",
+  "scores": {
+    "test": [
+      {
+        "hf_subset": "default",
+        "languages": [
+          "eng-Latn"
+        ],
+        "main_score": 0.42083101872374784,
+        "v_measure": 0.42083101872374784,
+        "v_measure_std": 0.1505123615549479,
+        "v_measures": [
+          0.3968523010343196,
+          0.3838234459188755,
+          0.37947436898429604,
+          0.35942070717275937,
+          0.40455386798014165,
+          0.4069209816716664,
+          0.3804144584425002,
+          0.4035132367607555,
+          0.41888301486092294,
+          0.38758976363961145,
+          0.5231988996608681,
+          0.5322423943945923,
+          0.5133751911298415,
+          0.5167438268299368,
+          0.5203429251878364,
+          0.5133760049134979,
+          0.5272652202379506,
+          0.5077927519152426,
+          0.5250978076493664,
+          0.5191940234807637,
+          0.4968698359778336,
+          0.2864511936403685,
+          0.30827031903814395,
+          0.40352359690958417,
+          0.32998636302856654,
+          0.22758595642345789,
+          0.28231167056783074,
+          0.1373717708120844,
+          0.22726891863130838,
+          1.0,
+          0.2260467635412604
+        ]
+      }
     ]
-  }
+  },
+  "task_name": "ArxivClusteringP2P.v2"
 }
