+++ results/sentence-transformers__paraphrase-multilingual-MiniLM-L12-v2/BlurbsClusteringP2P.v2.json
@@ -1,0 +1,50 @@
+{
+  "dataset_revision": "a2dd5b02a77de3466a3eaa98ae586b5610314496",
+  "evaluation_time": 309.5534794330597,
+  "kg_co2_emissions": null,
+  "mteb_version": "1.11.19",
+  "scores": {
+    "test": [
+      {
+        "hf_subset": "default",
+        "languages": [
+          "deu-Latn"
+        ],
+        "main_score": 0.3307291094644436,
+        "v_measure": 0.3307291094644436,
+        "v_measure_std": 0.11110904600828789,
+        "v_measures": [
+          0.22642808860283828,
+          0.24439552023334093,
+          0.22744895201615054,
+          0.229958943864066,
+          0.2107459790981925,
+          0.23216434060867086,
+          0.208008712762596,
+          0.24507944863624398,
+          0.23772322770328705,
+          0.2564261795242179,
+          0.3232343245235793,
+          0.5132794946572398,
+          0.2441159101261604,
+          0.22357822722978138,
+          0.20704054889853948,
+          0.2744465058837726,
+          0.2079605142537061,
+          0.5026109510843398,
+          0.44144927498664,
+          0.4595461659353289,
+          0.43698289291706255,
+          0.44483596277152476,
+          0.4600815173475022,
+          0.4477863685487717,
+          0.43822553025271427,
+          0.43177735724203387,
+          0.44573352306387953,
+          0.43935060223224137
+        ]
+      }
+    ]
+  },
+  "task_name": "BlurbsClusteringP2P.v2"
+}
