+++ results/sentence-transformers__paraphrase-multilingual-MiniLM-L12-v2/BlurbsClusteringS2S.v2.json
@@ -1,41 +1,50 @@
 {
   "dataset_revision": "22793b6a6465bf00120ad525e38c51210858132c",
-  "mteb_dataset_name": "BlurbsClusteringS2S.v2",
-  "mteb_version": "1.6.37",
-  "test": {
-    "evaluation_time": 156.9,
-    "main_score": 0.16963772029177934,
-    "v_measure": 0.16963772029177934,
-    "v_measure_std": 0.11073124051127854,
-    "v_measures": [
-      0.06807698245215885,
-      0.04455016075442265,
-      0.04642082953259332,
-      0.06198140475920798,
-      0.06317951799216749,
-      0.08015959222928008,
-      0.06660255770645218,
-      0.058429044939215316,
-      0.06850686845769428,
-      0.03850198082166569,
-      0.12555571434721985,
-      0.24353420865515152,
-      0.014695261667615252,
-      0.10619685444603377,
-      0.10204748367337636,
-      0.1658730460933846,
-      0.1358657548085803,
-      0.24818468133739777,
-      0.29780607460681713,
-      0.3076798811758003,
-      0.29106299539365793,
-      0.3048550767916618,
-      0.29954124288625006,
-      0.29953007386487546,
-      0.3179986822738043,
-      0.2990260265256237,
-      0.29156375208497726,
-      0.30243041789273567
+  "evaluation_time": 41.916991233825684,
+  "kg_co2_emissions": null,
+  "mteb_version": "1.11.19",
+  "scores": {
+    "test": [
+      {
+        "hf_subset": "default",
+        "languages": [
+          "deu-Latn"
+        ],
+        "main_score": 0.17009013783328794,
+        "v_measure": 0.17009013783328794,
+        "v_measure_std": 0.10750782282710497,
+        "v_measures": [
+          0.06913220796998218,
+          0.06243072235729465,
+          0.07719897492291466,
+          0.047366359848610255,
+          0.05089032746094627,
+          0.05575992018786144,
+          0.0772763659644507,
+          0.08751545545649228,
+          0.06773687116229814,
+          0.04949504627531629,
+          0.1273725355310106,
+          0.1996164446794968,
+          0.037943523609482104,
+          0.09752163085828966,
+          0.08754759692468538,
+          0.17210438992314073,
+          0.13247604284371758,
+          0.28707705497842806,
+          0.2959884722411192,
+          0.30562299444508706,
+          0.29794471889071894,
+          0.2956063139304293,
+          0.298852991309102,
+          0.2882823520441266,
+          0.3118354828580177,
+          0.2959372783536503,
+          0.28741034130661414,
+          0.2985814429987785
+        ]
+      }
     ]
-  }
+  },
+  "task_name": "BlurbsClusteringS2S.v2"
 }
