+++ results/sentence-transformers__paraphrase-multilingual-MiniLM-L12-v2/TenKGnadClusteringP2P.v2.json
@@ -1,0 +1,33 @@
+{
+  "dataset_revision": "5c59e41555244b7e45c9a6be2d720ab4bafae558",
+  "evaluation_time": 66.06710958480835,
+  "kg_co2_emissions": null,
+  "mteb_version": "1.11.19",
+  "scores": {
+    "test": [
+      {
+        "hf_subset": "default",
+        "languages": [
+          "deu-Latn"
+        ],
+        "main_score": 0.37046433294765324,
+        "v_measure": 0.37046433294765324,
+        "v_measures": {
+          "Level 0": [
+            0.3728441063790658,
+            0.36486251369389927,
+            0.36543828858662325,
+            0.35938174052697475,
+            0.3698364255367986,
+            0.4349418143287204,
+            0.33869628379037836,
+            0.3970068872033539,
+            0.36956197486272174,
+            0.33207329456799656
+          ]
+        }
+      }
+    ]
+  },
+  "task_name": "TenKGnadClusteringP2P.v2"
+}
