+++ results/sentence-transformers__paraphrase-multilingual-MiniLM-L12-v2/TenKGnadClusteringS2S.v2.json
@@ -1,21 +1,33 @@
 {
   "dataset_revision": "6cddbe003f12b9b140aec477b583ac4191f01786",
-  "mteb_dataset_name": "TenKGnadClusteringS2S.v2",
-  "mteb_version": "1.6.37",
-  "test": {
-    "evaluation_time": 35.54,
-    "v_measure": 0.21597117519062153,
-    "v_measures": [
-      0.20303783002621278,
-      0.2133904852293011,
-      0.2421962190466207,
-      0.23723125738439813,
-      0.2058062231396479,
-      0.20574773195326623,
-      0.20069622544566787,
-      0.21638662343636447,
-      0.2208966644188706,
-      0.21432249182586569
+  "evaluation_time": 12.912961959838867,
+  "kg_co2_emissions": null,
+  "mteb_version": "1.11.19",
+  "scores": {
+    "test": [
+      {
+        "hf_subset": "default",
+        "languages": [
+          "deu-Latn"
+        ],
+        "main_score": 0.2139429095585484,
+        "v_measure": 0.2139429095585484,
+        "v_measures": {
+          "Level 0": [
+            0.24314744588777754,
+            0.2011998687454228,
+            0.23513748984026547,
+            0.21287863717886216,
+            0.23435525912240007,
+            0.17524914303574923,
+            0.22894648960071265,
+            0.20953803255162262,
+            0.1814406916592189,
+            0.21753603796345267
+          ]
+        }
+      }
     ]
-  }
+  },
+  "task_name": "TenKGnadClusteringS2S.v2"
 }
